Fixes for `kill_pid`, `interrupt`, and `log` as well as tests.
The previous patch to make `interrupt` cleaner was poorly thought out.
`interrupt` will now only raise if :debug is set, otherwise it passes a
logged message.

`kill_pid` now uses SIGTERM instead of SIGINT since stopping a daemon
cleanly isn't actually the same as interrupting a running process from
the terminal; SIGTERM is more appropriate than SIGINT in that context.

After changing the handling of SIGTERM in runner.rb, the test
needed to be updated to accommodate the fact that SIGTERM is no
longer processed as SIGINT. I've also rolled in a test to validate that
SIGINT and SIGTERM are being handled differently. All tests now pass.

`log` has also been changed so that all messages are logged, regardless of
debug mode. This is more in-line with how C daemons work. I think that it
could be refactored to use Logger but for now it will provide a relatively
reliable audit trail for applications flapping.